requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
backports.entry-points-selectable==1.1.1
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.9
click==8.0.3
dataclasses==0.6
distlib==0.3.4
filelock==3.4.0
Flask==2.0.2
frozenlist==1.2.0
future==0.18.2
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
gspread==5.0.0
httplib2==0.20.2
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
line-bot-sdk==2.0.1
MarkupSafe==2.0.1
multidict==5.2.0
oauthlib==3.1.1
pipenv==2021.11.23
platformdirs==2.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygraphviz==1.3
pygsheets==2.0.5
pyparsing==3.0.6
PySheets==1.0.3
python-dotenv==0.19.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
transitions==0.8.10
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
utils==1.0.1
virtualenv==20.10.0
virtualenv-clone==0.5.7
Werkzeug==2.0.2
yarl==1.7.2
zipp==3.6.0
psycopg2==2.6.1
